make save

+++ Fdf_2/includes/fdf_struct.h
@@ -6,7 +6,7 @@
 /*   By: qbanet <qbanet@student.42perpignan.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/31 11:41:22 by qbanet            #+#    #+#             */
-/*   Updated: 2023/08/16 08:04:36 by qbanet           ###   ########.fr       */
+/*   Updated: 2023/08/16 10:24:45 by qbanet           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,12 +63,12 @@ typedef struct s_line
 	t_color				colors;
 	t_point				start;
 	t_point				end;
-	int					dx;
-	int					dy;
+	float				dx;
+	float				dy;
 	int					sx;
 	int					sy;
-	int					err;
-	int					err2;
+	float				err;
+	float				err2;
 }	t_line;
 
 typedef struct s_img
